Adds support for TDX tests
Uses computeBeta API to check for an image's GuestOsFeatures instead of the Compute API since TDX_CAPABLE feature is only listed with computeBeta.

It also changes the machine type for the TDX tests from c3-standard-2 to c3-standard-4. This is because the c3-standard-2 machine type does not support the TDX feature.

Finally, this CL also updates the daisy compute library to include a new method called GetImageFromFamilyBeta. This method gets an image from an image family using the Beta API.

PiperOrigin-RevId: 650540484