Fix imageboot test to handle a couple of error conditions
In Rocky 8, when a reboot command is sent, the connection is often
terminated before being properly disconnected, which is read as an error in
this test.  By checking whether the error message contains "signal:
terminated" and only raising an error if it doesn't, we ensure that the
test gets past this step.

There is a second problem where a good number of the instance types don't
reboot quickly enough to kill the test before the fatal error "marker file
does not exist" is triggered.  By adding a one minute sleep, we ensure that
the fatal error is only triggered if the guest hasn't rebooted within 60
seconds.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>