Generalize HyperdiskNeeded function so it's used everywhere it's required
We have a number of places where we check if a hyperdisk is required for a
machine type, but each check is slightly different.  In the last commit, a
new function, `HyperdiskNeeded` was added to the hotattach test that has
the most comprehensive check on whether or not a hyperdisk is required.

This commit moves that function to utils/test_utils.go and updates all the
other places where we need a hyperdisk to use that function, giving us one
source of truth as to whether or not a hyperdisk is required for a machine
type.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>