Copybara import of the project:
--
e5cf7390c9e23400941ded9ce96ddff5f6aa5381 by Jonathan Dieter <jdieter@ciq.com>:

Generalize HyperdiskNeeded function so it's used everywhere it's required

We have a number of places where we check if a hyperdisk is required for a
machine type, but each check is slightly different.  In the last commit, a
new function, `HyperdiskNeeded` was added to the hotattach test that has
the most comprehensive check on whether or not a hyperdisk is required.

This commit moves that function to utils/test_utils.go and updates all the
other places where we need a hyperdisk to use that function, giving us one
source of truth as to whether or not a hyperdisk is required for a machine
type.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>

--
75218f69d6a4e8beebd6966ba3cdd058f0f5757c by Jonathan Dieter <jdieter@ciq.com>:

Fix imageboot test to handle a couple of error conditions

In Rocky 8, when a reboot command is sent, the connection is often
terminated before being properly disconnected, which is read as an error in
this test.  By checking whether the error message contains "signal:
terminated" and only raising an error if it doesn't, we ensure that the
test gets past this step.

There is a second problem where a good number of the instance types don't
reboot quickly enough to kill the test before the fatal error "marker file
does not exist" is triggered.  By adding a one minute sleep, we ensure that
the fatal error is only triggered if the guest hasn't rebooted within 60
seconds.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>
FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/GoogleCloudPlatform/cloud-image-tests/pull/243 from jdieter:test-fixes 75218f69d6a4e8beebd6966ba3cdd058f0f5757c
PiperOrigin-RevId: 783927210