Fix web server so it closes connections before exiting
We were hitting race conditions where the data meant to be written out in
http.HandleFunc was not being received by the client, breaking the tests.

The problem is that srv.ListenAndServe() unblocks immediately when
srv.Shutdown() is called, and, HandleFunc doesn't close the request on
exit, which means the mutex doesn't actually protect anything.

By moving ListenAndServe into the goroutine, and Shutdown into the main
function, we don't actually exit the main function until Shutdown is
complete, allowing all requests to complete before the main function exits.

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>