Add support for running tests across multiple zones.
Introduce a new `-zones` flag to allow specifying a comma-separated list of zones. Tests will be distributed across these zones in a round-robin fashion. The single `-zone` flag is still supported for running all tests in a single zone.

The zone parameter has been removed from `finalizeWorkflows` and related functions as the zone is now determined per test suite in and it was previously overriding the previously prepared/preset structure.

PiperOrigin-RevId: 792707101