Reduce retry attempts and delays in nicsetup test suite.
When making connections, with the previously set timeout values, we can wait up to 5 minutes to make a connection. However, there's no need to wait for that long, or make as many connection attempts, since by the time we attempt to connect, the receiver of these connections should already be available for connections, and if the sender is not meant to send the traffic, then it will be able to fail faster.

Reducing the timeout should reduce the average runtime by...about 5 minutes.
Also add some comments around the constants and some reasonings behind the values set.

PiperOrigin-RevId: 797041978