Move all the metadata operations in the beginning of the test, and consolidate the primary and secondary NIC tests.
For some reason, at least on Debian 12, if you try to do an HTTP request to MDS directly after restarting the guest agent, the request hangs for up to 15 minutes. Moving all the MDS calls to before the guest agent restarts and testing should help avoid that issue.

We also add a timeout to the HTTP client for MDS requests to avoid indefinite hangs.

PiperOrigin-RevId: 799770385