Add WSFC health check agent tests to cloud\_image\_tests.
This change introduces a new test suite `wsfc` to validate the functionality of the Windows Server Failover Cluster health check agent. The test suite includes `TestHealthCheckAgent`, which sends TCP requests to the agent on a specific port and verifies the responses based on whether the provided IP address is valid or not. The test setup enables the WSFC agent via VM metadata.

PiperOrigin-RevId: 816914598